finished implementing email support (client side also ready and successfully locally tested)
